UI_LogIn
This application in made by using cordova in ionic frame work. This
version  has User interface with log in and home page for sports meet
up app. The UI part is working in local host and iOS emulater.  Merging
of backend with database is yet to be done. Once this merging is done
we can associate this component with UI.